Add files via upload
This file showcases my Excel skills and the ability to interpret, transform and then visualize that data. By using the customer and sales data I was able to create a Histogram with 5 bins and a bin width of 482. I also created a XY scatter plot chart with a linear trend line to make it easier to visualize the relationship of the positive correlation between the amount of customers and sales. I also created a running average to keep track as average sales changed on a daily basis.