Add files via upload
By using data I had on distance traveled to a store as well as the amount spent based on the distance traveled, I was able to forecast how much an individual driving 30 miles would spend based on historical data. Using the data, I created a linear forecast formula, specifying x, known ys and known xs. This was how I was able to forecast what an individual may spend if they were traveling a distance of 30 miles based on historical data.